Update flow for wallet list sheet. Add design from cosmos_ui_components for wallet list sheet

+++ lib/data/model/token_json.dart
@@ -1,7 +1,7 @@
 class TokenJson {
   final String symbol;
   final double price;
-  final int supply;
+  final double supply;
 
   const TokenJson({required this.symbol, required this.price, required this.supply});
 
@@ -9,7 +9,7 @@ class TokenJson {
     return TokenJson(
       symbol: json['Symbol'] as String? ?? '',
       price: json['Price'] as double? ?? 0.0,
-      supply: json['Supply'] as int? ?? -1,
+      supply: json['Supply'] as double? ?? -1,
     );
   }
 }
