+++ lib/domain/entities/token.dart
@@ -4,7 +4,7 @@ import 'package:flutter_app/domain/entities/denom.dart';
 class Token {
   late Denom denom;
   late Amount amount;
-  late int supply;
+  late double supply;
 
   Token({required this.denom, required this.amount, required this.supply});
 }
